--- frontend/src/components/PluginDetails/index.ts
@@ -1,1 +1,0 @@
-export * from './PluginDetails';
