+++ frontend/src/components/PluginPage/CitationInfo.constants.ts
