+++ frontend/src/components/PluginPage/CitationInfo.tsx
