+++ frontend/src/components/PluginPage/InstallModal.module.scss
