+++ frontend/src/components/PluginPage/InstallModal.module.scss.d.ts
