+++ frontend/src/components/PluginPage/InstallModal.tsx
