+++ frontend/src/components/PluginPage/PluginMetadata.tsx
