+++ frontend/src/components/PluginPage/SupportInfo.module.scss.d.ts
