+++ frontend/src/components/PluginPage/index.ts
@@ -1,0 +1,1 @@
+export * from './PluginPage';
